scan nacos style
