+++ common/src/main/java/com/alibaba/nacos/common/utils/ReflectUtils.java
@@ -17,6 +17,9 @@
 package com.alibaba.nacos.common.utils;
 
 import java.lang.reflect.Field;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.lang.reflect.UndeclaredThrowableException;
 
 /**
  * reflect utils.
@@ -59,5 +62,113 @@ public class ReflectUtils {
             return defaultValue;
         }
     }
-    
+
+    /**
+     * Get the field represented by the supplied {@link Field field object} on the
+     * specified {@link Object target object}. In accordance with {@link Field#get(Object)}
+     * semantics, the returned value is automatically wrapped if the underlying field
+     * has a primitive type.
+     *
+     * <p>Thrown exceptions are handled via a call to {@link #handleReflectionException(Exception)}.
+     *
+     * @param field  the field to get
+     * @param target the target object from which to get the field
+     *               (or {@code null} for a static field)
+     * @return the field's current value
+     */
+    public static Object getField(Field field, Object target) {
+        try {
+            return field.get(target);
+        } catch (IllegalAccessException ex) {
+            handleReflectionException(ex);
+        }
+        throw new IllegalStateException("Should never get here");
+    }
+
+    /**
+     * Handle the given reflection exception.
+     *
+     * <p>Should only be called if no checked exception is expected to be thrown
+     * by a target method, or if an error occurs while accessing a method or field.
+     *
+     * <p>Throws the underlying RuntimeException or Error in case of an
+     * InvocationTargetException with such a root cause. Throws an
+     * IllegalStateException with an appropriate message or
+     * UndeclaredThrowableException otherwise.
+     *
+     * @param ex the reflection exception to handle
+     */
+    public static void handleReflectionException(Exception ex) {
+        if (ex instanceof NoSuchMethodException) {
+            throw new IllegalStateException("Method not found: " + ex.getMessage());
+        }
+        if (ex instanceof IllegalAccessException) {
+            throw new IllegalStateException("Could not access method or field: " + ex.getMessage());
+        }
+        if (ex instanceof InvocationTargetException) {
+            handleInvocationTargetException((InvocationTargetException) ex);
+        }
+        if (ex instanceof RuntimeException) {
+            throw (RuntimeException) ex;
+        }
+        throw new UndeclaredThrowableException(ex);
+    }
+
+    /**
+     * Handle the given invocation target exception. Should only be called if no
+     * checked exception is expected to be thrown by the target method.
+     *
+     * <p>Throws the underlying RuntimeException or Error in case of such a root
+     * cause. Throws an UndeclaredThrowableException otherwise.
+     *
+     * @param ex the invocation target exception to handle
+     */
+    public static void handleInvocationTargetException(InvocationTargetException ex) {
+        rethrowRuntimeException(ex.getTargetException());
+    }
+
+    /**
+     * Rethrow the given {@link Throwable exception}, which is presumably the
+     * <em>target exception</em> of an {@link InvocationTargetException}.
+     * Should only be called if no checked exception is expected to be thrown
+     * by the target method.
+     *
+     * <p>Rethrows the underlying exception cast to a {@link RuntimeException} or
+     * {@link Error} if appropriate; otherwise, throws an
+     * {@link UndeclaredThrowableException}.
+     *
+     * @param ex the exception to rethrow
+     * @throws RuntimeException the rethrown exception
+     */
+    public static void rethrowRuntimeException(Throwable ex) {
+        if (ex instanceof RuntimeException) {
+            throw (RuntimeException) ex;
+        }
+        if (ex instanceof Error) {
+            throw (Error) ex;
+        }
+        throw new UndeclaredThrowableException(ex);
+    }
+
+    /**
+     * Invoke the specified {@link Method} against the supplied target object with the
+     * supplied arguments. The target object can be {@code null} when invoking a
+     * static {@link Method}.
+     *
+     * <p>Thrown exceptions are handled via a call to {@link #handleReflectionException}.
+     *
+     * @param method the method to invoke
+     * @param target the target object to invoke the method on
+     * @param args   the invocation arguments (may be {@code null})
+     * @return the invocation result, if any
+     */
+    public static Object invokeMethod(Method method, Object target, Object... args) {
+        try {
+            return method.invoke(target, args);
+        } catch (Exception ex) {
+            handleReflectionException(ex);
+        }
+        throw new IllegalStateException("Should never get here");
+    }
+
 }
